pyproject.toml
[project]
name = "dietai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # FastAPI和Web服务
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "sse-starlette>=1.6.5",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.3",
    
    # 数据库和ORM
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    
    # 认证和安全
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "email-validator>=2.1.0",
    
    # 缓存和存储
    "redis>=5.0.1",
    "minio>=7.2.0",
    
    # 数据处理
    "pandas>=2.2.3",
    "numpy>=2.2.3",
    "pillow>=10.0.0",
    "python-slugify>=8.0.1",
    "aiofiles>=23.2.1",
    
    # 配置和环境
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.1",
    "pytz>=2023.3",
    
    # HTTP客户端
    "httpx>=0.28.1",
    "requests>=2.31.0",
    
    # 测试
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    
    # LangChain相关（保留AI功能）
    "langchain-community>=0.3.19",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.8",
    "langgraph>=0.3.5",
    "langchain-anthropic>=0.3.16",
    "langchain-qwq>=0.2.0",
    "dashscope>=1.23.6",
    "langchain-mcp-adapters>=0.0.9",
    "langgraph-cli[inmem]>=0.2.10",
    "mcp>=1.6.0",
    "litellm>=1.63.11",
    
    # 工具和实用
    "readabilipy>=0.3.0",
    "socksio>=1.0.0",
    "markdownify>=1.1.0",
    "json-repair>=0.7.0",
    "duckduckgo-search>=8.0.0",
    "inquirerpy>=0.3.4",
    "arxiv>=2.2.0",
    "yfinance>=0.2.54",
    "jupyter>=1.1.1",
    "croniter>=1.4.1",
    "celery>=5.3.0",
    "flower>=2.0.1",
]
